Change most usize to u64

+++ sonance/2/src/vm/frames.rs
@@ -4,7 +4,7 @@ use crate::error::{Error, Result};
 
 #[derive(Debug, Clone, Default, PartialEq)]
 pub struct Frame {
-    pub variables: HashMap<usize, usize>,
+    pub variables: HashMap<u64, u64>,
     pub return_index: usize,
 }
 
@@ -38,13 +38,13 @@ impl Frames {
         Ok(frame.return_index)
     }
 
-    pub fn load(&self, variable: usize) -> Result<usize> {
+    pub fn load(&self, variable: u64) -> Result<u64> {
         let frame = self.frames.last().ok_or(Error::ExpectedFrame)?;
 
         Ok(frame.variables.get(&variable).copied().unwrap_or(0))
     }
 
-    pub fn store(&mut self, variable: usize, value: usize) -> Result<()> {
+    pub fn store(&mut self, variable: u64, value: u64) -> Result<()> {
         let frame = self.frames.last_mut().ok_or(Error::ExpectedFrame)?;
 
         frame.variables.insert(variable, value);
